requirements.txt
albumentations==1.1.0
ipykernel==5.3.4
ipython==7.22.0
jupyter==1.0.0
matplotlib==3.3.4
pandas==1.2.4
Pillow==8.2.0
opencv-python==4.5.3.56
scikit-learn==0.24.1
scikit-image==0.18.1
scipy==1.6.2
seaborn==0.11.1
torch==1.9.0
torchvision==0.10.0
torchaudio==0.9.0



